продолжаю разносить компоненты по файлам и папкам
