+++ src/view/films-list-empty/films-list-empty-template.js
@@ -1,0 +1,5 @@
+export const createFilmsListEmptyTemplate = () => {
+  return `<section class="films-list">
+    <h2 class="films-list__title">There are no movies in our database</h2>
+  </section>`;
+};
