+++ src/view/films/films-template.js
@@ -1,0 +1,4 @@
+export const createFilmsWrapperTemplate = () => {
+  return `<section class="films">
+    </section>`;
+};
