+++ src/view/profile/profile-template.js
@@ -1,5 +1,3 @@
-import {createElement} from "../utils";
-
 const Ranks = {
   NOVICE: `Novice`,
   NOVICE_FILMS_MAX: 10,
@@ -22,32 +20,9 @@ const getRank = (fimlsCount) => {
   return rank;
 };
 
-const createProfileTemplate = (fimlsCount) => {
+export const createProfileTemplate = (fimlsCount) => {
   return `<section class="header__profile profile">
     <p class="profile__rating">${getRank(fimlsCount)}</p>
     <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
   </section>`;
 };
-
-export default class Profile {
-  constructor(filmsCount) {
-    this._element = null;
-    this._filmsCount = filmsCount;
-  }
-
-  getTemplate() {
-    return createProfileTemplate(this._filmsCount);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
