+++ src/view/show-more-button/show-more-button-template.js
@@ -1,0 +1,3 @@
+export const createShowMoreButtonTemplate = () => {
+  return `<button class="films-list__show-more">Show more</button>`;
+};
