+++ src/view/site-menu/site-menu-template.js
@@ -1,6 +1,4 @@
-import {createElement} from "../utils";
-
-const createSiteMenuTemplate = (cards) => {
+export const createSiteMenuTemplate = (cards) => {
   return `<nav class="main-navigation">
     <div class="main-navigation__items">
       <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
@@ -11,26 +9,3 @@ const createSiteMenuTemplate = (cards) => {
     <a href="#stats" class="main-navigation__additional">Stats</a>
   </nav>`;
 };
-
-export default class SiteMenu {
-  constructor(cards) {
-    this._element = null;
-    this._cards = cards;
-  }
-
-  getTemplate() {
-    return createSiteMenuTemplate(this._cards);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
